Move Violations out of the internal package since they are part of the public api
