+++ plugin/src/main/groovy/com/novoda/staticanalysis/ViolationsEvaluator.groovy
@@ -1,7 +1,5 @@
 package com.novoda.staticanalysis
 
-import com.novoda.staticanalysis.internal.Violations
-
 interface ViolationsEvaluator {
 
     void evaluate(Set<Violations> allViolations)
