+++ plugin/src/main/groovy/com/novoda/staticanalysis/internal/CollectViolationsTask.groovy
@@ -1,5 +1,6 @@
 package com.novoda.staticanalysis.internal
 
+import com.novoda.staticanalysis.Violations
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
